+++ cfme/infrastructure/provider/__init__.py
@@ -10,7 +10,7 @@ from widgetastic_patternfly import BreadCrumb
 from cfme.base.ui import Server
 from cfme.common import TagPageView, Taggable, PolicyProfileAssignable
 from cfme.common.host_views import ProviderAllHostsView
-from cfme.common.provider import BaseProvider, provider_types
+from cfme.common.provider import BaseProvider, provider_types, CloudInfraProviderMixin
 from cfme.common.provider_views import (
     InfraProviderAddView, InfraProviderEditView, InfraProviderDetailsView, ProviderTimelinesView,
     InfraProvidersDiscoverView, InfraProvidersView, ProviderNodesView, ProviderTemplatesView,
@@ -22,7 +22,6 @@ from cfme.infrastructure.virtual_machines import InfraVm, InfraTemplate
 from cfme.modeling.base import BaseCollection
 from cfme.utils.appliance.implementations.ui import navigator, CFMENavigateStep, navigate_to
 from cfme.utils.log import logger
-from cfme.utils.net import resolve_hostname
 from cfme.utils.pretty import Pretty
 from cfme.utils.varmeth import variable
 from cfme.utils.wait import wait_for
@@ -44,7 +43,8 @@ class ProviderClustersView(ClusterView):
 
 
 @attr.s(hash=False)
-class InfraProvider(Pretty, Fillable, BaseProvider, PolicyProfileAssignable, Taggable):
+class InfraProvider(BaseProvider, CloudInfraProviderMixin, Pretty, Fillable,
+                    PolicyProfileAssignable, Taggable):
     """
     Abstract model of an infrastructure provider in cfme. See VMwareProvider or RHEVMProvider.
 
@@ -77,9 +77,6 @@ class InfraProvider(Pretty, Fillable, BaseProvider, PolicyProfileAssignable, Tag
     db_types = ["InfraManager"]
     hosts_menu_item = "Hosts"
     vm_name = "Virtual Machines"
-    detail_page_suffix = 'provider'
-    edit_page_suffix = 'provider_edit'
-    refresh_text = "Refresh Relationships and Power States"
 
     name = attr.ib(default=None)
     key = attr.ib(default=None)
@@ -93,91 +90,6 @@ class InfraProvider(Pretty, Fillable, BaseProvider, PolicyProfileAssignable, Tag
     def __attrs_post_init__(self):
         super(InfraProvider, self).__attrs_post_init__()
         self.parent = self.appliance.collections.infra_providers
-
-    @property
-    def hostname(self):
-        return getattr(self.default_endpoint, "hostname", None)
-
-    @hostname.setter
-    def hostname(self, value):
-        if self.default_endpoint:
-            if value:
-                self.default_endpoint.hostname = value
-        else:
-            logger.warn("can't set hostname because default endpoint is absent")
-
-    @property
-    def ip_address(self):
-        return getattr(self.default_endpoint, "ipaddress", resolve_hostname(str(self.hostname)))
-
-    @ip_address.setter
-    def ip_address(self, value):
-        if self.default_endpoint:
-            if value:
-                self.default_endpoint.ipaddress = value
-        else:
-            logger.warn("can't set ipaddress because default endpoint is absent")
-
-    @variable(alias="db")
-    def num_template(self):
-        """ Returns the providers number of templates, as shown on the Details page."""
-        ext_management_systems = self.appliance.db.client["ext_management_systems"]
-        vms = self.appliance.db.client["vms"]
-        temlist = list(self.appliance.db.client.session.query(vms.name)
-                       .join(ext_management_systems, vms.ems_id == ext_management_systems.id)
-                       .filter(ext_management_systems.name == self.name)
-                       .filter(vms.template == True))  # NOQA
-        return len(temlist)
-
-    @num_template.variant('ui')
-    def num_template_ui(self):
-        view = navigate_to(self, "Details")
-        return int(view.entities.summary("Relationships").get_text_of(self.template_name))
-
-    @variable(alias="db")
-    def num_vm(self):
-        """ Returns the providers number of instances, as shown on the Details page."""
-        ext_management_systems = self.appliance.db.client["ext_management_systems"]
-        vms = self.appliance.db.client["vms"]
-        vmlist = list(self.appliance.db.client.session.query(vms.name)
-                      .join(ext_management_systems, vms.ems_id == ext_management_systems.id)
-                      .filter(ext_management_systems.name == self.name)
-                      .filter(vms.template == False))  # NOQA
-        return len(vmlist)
-
-    @num_vm.variant('ui')
-    def num_vm_ui(self):
-        view = navigate_to(self, "Details")
-        return int(view.entities.summary("Relationships").get_text_of(self.vm_name))
-
-    def load_all_provider_instances(self):
-        return self.load_all_provider_vms()
-
-    def load_all_provider_vms(self):
-        """ Loads the list of instances that are running under the provider.
-
-        """
-        view = navigate_to(self, 'Details')
-        if view.entities.summary("Relationships").get_text_of(self.vm_name) == "0":
-            return False
-        else:
-            view.entities.summary("Relationships").click_at(self.vm_name)
-            return True
-
-    def load_all_provider_images(self):
-        self.load_all_provider_templates()
-
-    def load_all_provider_templates(self):
-        """ Loads the list of images that are available under the provider.
-
-        """
-        # todo: replace these methods with new nav location
-        view = navigate_to(self, 'Details')
-        if view.entities.summary("Relationships").get_text_of(self.template_name) == "0":
-            return False
-        else:
-            view.entities.summary("Relationships").click_at(self.template_name)
-            return True
 
     @variable(alias='db')
     def num_datastore(self):
